requirements.txt
fastapi 
uvicorn 
jinja2 
requests 
python-multipart

openai 
python-dotenv 
markdown
opik
